// referral-form-data-process/lib/interfaces/residentFormData.ts
export interface ResidentFormDataObject {
  firstName: string | undefined;
  lastName: string | undefined;
  dateOfBirth: string | undefined;
  gender: string | undefined;
  ethnicity: string | undefined;
  firstLanguage: string | undefined;
  school: string | undefined;
  addressLineOne: string | undefined;
  addressLineTwo: string | undefined;
  addressLineThree: string | undefined;
  postcode: string | undefined;
}
export interface MashResidentRequest {
  firstName: string | undefined;
  lastName: string | undefined;
  dateOfBirth: string | undefined;
  gender: string | undefined;
  ethnicity: string | undefined;
  firstLanguage: string | undefined;
  school: string | undefined;
  address: string | undefined;
  postcode: string | undefined;
}